updated design

+++ app/controllers/demo_controller.rb
@@ -1,42 +1,40 @@
 class DemoController < ApplicationController
   def create
     url = params[:url]
+    keyword = params[:keyword]
+    return redirect_to root_path unless url == 'google'
 
-require 'headless'
-headless = Headless.new(dimensions: "1400x900x24", video: { frame_rate: 12, codec: 'libx264', provider: 'ffmpeg' })
+    require 'headless'
+    headless = Headless.new(dimensions: "1400x900x24", video: { frame_rate: 12, codec: 'libx264', provider: 'ffmpeg' })
 
-# headless = Headless.new(dimensions: "1400x900x24")
-puts headless
-puts headless.display
-puts headless.dimensions
-headless.start
+    # headless = Headless.new(dimensions: "1400x900x24")
+    headless.start
 
-#Before do
-  headless.video.start_capture
-#end
+    headless.video.start_capture
 
-driver = Selenium::WebDriver.for :chrome
+    driver = Selenium::WebDriver.for :chrome
 
-driver.navigate.to "http://google.com"
+    driver.navigate.to "https://www.google.com"
 
-element = driver.find_element(:name, 'q')
-element.send_keys "Hello RocketQA"
-# element.submit
-driver.find_element(:class, "lsb").click
+    element = driver.find_element(:name, 'q')
+    element.send_keys(keyword || "Hello RocketQA")
+    # element.submit
+    driver.find_element(:class, "lsb").click
 
-driver.save_screenshot('/home/ubuntu/projects/RocketQA/public/sample.png')
-puts driver.title
+    driver.save_screenshot('/home/ubuntu/projects/RocketQA/public/sample.png')
 
-#After do |scenario|
-#  if scenario.failed?
-    headless.video.stop_and_save("/home/ubuntu/projects/RocketQA/public/sample.mp4")
-#  else
-#    headless.video.stop_and_discard
-#  end
-#end
+    #After do |scenario|
+    #  if scenario.failed?
+        headless.video.stop_and_save("/home/ubuntu/projects/RocketQA/public/sample.mp4")
+    #  else
+    #    headless.video.stop_and_discard
+    #  end
+    #end
 
-headless.destroy
+    headless.destroy
 
+    redirect_to root_path
+  rescue
     redirect_to root_path
   end
 end
